# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GET IT NOW API",
    "version": "1.0.0",
    "description": "On-demand service marketplace API for connecting customers with verified workers",
    "contact": {
      "name": "API Support",
      "email": "support@getitnow.example"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.getitnow.example",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/reports/bookings": {
      "get": {
        "summary": "Get bookings report",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "serviceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cooperativeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bookings report"
          }
        }
      }
    },
...
    "/payments/invoices/booking/{bookingId}": {
      "get": {
        "summary": "Get invoice by booking",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details"
          }
        }
      }
    },
...
    "/matching/bookings/{bookingId}/reassign": {
      "post": {
        "summary": "Reassign a booking to a different worker",
        "tags": [
          "Matching"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "workerId"
                ],
                "properties": {
                  "workerId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Booking reassigned successfully"
          },
          "404": {
            "description": "Booking or worker not found"
          },
          "409": {
            "description": "Worker not available"
          }
        }
      }
    },
...
    "/institutions/{id}/bookings": {
      "get": {
        "summary": "List organization bookings",
        "tags": [
          "Institutions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings"
          }
        }
      }
    },
...
    "/emergency/bookings": {
      "post": {
        "summary": "Create emergency booking",
        "tags": [
          "Emergency"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "serviceId",
                  "description",
                  "latitude",
                  "longitude",
                  "address"
                ],
                "properties": {
                  "serviceId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "description": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "address": {
                    "type": "string"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "critical",
                      "high",
                      "standard"
                    ]
                  },
                  "duplicateKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emergency booking created"
          },
          "409": {
            "description": "Duplicate emergency request"
          }
        }
      }
    },
...
    "/bookings": {
      "post": {
        "summary": "Create a booking",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "List the caller's bookings",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}": {
      "get": {
        "summary": "Get an authorized booking",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/status": {
      "patch": {
        "summary": "Transition a booking status",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/cancel": {
      "post": {
        "summary": "Cancel a booking",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/reschedule": {
      "post": {
        "summary": "Reschedule a booking",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/accept": {
      "post": {
        "summary": "Accept a booking (worker)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/reject": {
      "post": {
        "summary": "Reject a booking (worker)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/start": {
      "post": {
        "summary": "Start a booking (worker)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/complete": {
      "post": {
        "summary": "Complete a booking (worker)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/reassign": {
      "post": {
        "summary": "Reassign a booking (admin)",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/bookings/{id}/timeline": {
      "get": {
        "summary": "Get booking status timeline",
        "tags": [
          "Bookings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/analytics/bookings": {
      "get": {
        "summary": "Get bookings analytics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "serviceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cooperativeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "area",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bookings analytics"
          }
        }
      }
    },
...
  }
}